feat: expand Complete Saga support and audio bridging
- Add AudioQueue compatibility and broaden Core Foundation, controller, and event handling
- Improve Complete Saga gameplay, media playback, frame pacing, and hitch diagnostics
- Add focused bridge coverage and update build and runtime documentation

+++ native/src/mode32_probe.c
@@ -12,7 +12,7 @@
 #include <sys/sysctl.h>
 #include <unistd.h>
 
-extern uint32_t run_compat32(uint32_t eip, uint32_t esp, uint16_t cs32);
+extern uint64_t run_compat32(uint32_t eip, uint32_t esp, uint16_t cs32);
 
 enum {
     kCodeMapSize = 0x4000,
